feat: Redirs & heredoc routines change
Hope it works and I have no leaks. The idea is to ask every heredoc
asked and write it to a temp file (the same for all but wiped each
time). Then, at fork if a redir_in is asked, close the old open fd if it
was open (meaning the temp file of heredoc for 1st infile). Each redir_out
will close old openned file & open the new one. Same for redir_out.